Feature/item crud

+++ src/main/java/com/example/refrigerator/repository/RefrigeratorRepository.java
@@ -3,5 +3,5 @@ package com.example.refrigerator.repository;
 import com.example.refrigerator.domain.Refrigerator;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RefrigeratorRepository extends JpaRepository<Refrigerator, Integer> {
+public interface RefrigeratorRepository extends JpaRepository<Refrigerator, Long> {
 }
